+++ administrator/components/com_crowdfunding/views/locations/view.html.php
@@ -35,14 +35,9 @@ class CrowdfundingViewLocations extends JViewLegacy
 
     protected $sidebar;
 
-    public function __construct($config)
-    {
-        parent::__construct($config);
-        $this->option = JFactory::getApplication()->input->get("option");
-    }
-
     public function display($tpl = null)
     {
+        $this->option     = JFactory::getApplication()->input->get('option');
         $this->state      = $this->get('State');
         $this->items      = $this->get('Items');
         $this->pagination = $this->get('Pagination');
@@ -69,7 +64,7 @@ class CrowdfundingViewLocations extends JViewLegacy
         // Prepare filters
         $this->listOrder = $this->escape($this->state->get('list.ordering'));
         $this->listDirn  = $this->escape($this->state->get('list.direction'));
-        $this->saveOrder = (strcmp($this->listOrder, 'a.ordering') != 0) ? false : true;
+        $this->saveOrder = (strcmp($this->listOrder, 'a.ordering') === 0);
 
         if ($this->saveOrder) {
             $this->saveOrderingUrl = 'index.php?option=' . $this->option . '&task=' . $this->getName() . '.saveOrderAjax&format=raw';
@@ -119,25 +114,25 @@ class CrowdfundingViewLocations extends JViewLegacy
 
         // Import
         $link = JRoute::_('index.php?option=com_crowdfunding&view=import&type=locations');
-        $bar->appendButton('Link', 'upload', JText::_("COM_CROWDFUNDING_IMPORT_LOCATIONS"), $link);
+        $bar->appendButton('Link', 'upload', JText::_('COM_CROWDFUNDING_IMPORT_LOCATIONS'), $link);
 
         $link = JRoute::_('index.php?option=com_crowdfunding&view=import&type=states');
-        $bar->appendButton('Link', 'upload', JText::_("COM_CROWDFUNDING_IMPORT_STATES"), $link);
+        $bar->appendButton('Link', 'upload', JText::_('COM_CROWDFUNDING_IMPORT_STATES'), $link);
 
         // Export
         $link = JRoute::_('index.php?option=com_crowdfunding&task=export.download&format=raw&type=locations');
-        $bar->appendButton('Link', 'download', JText::_("COM_CROWDFUNDING_EXPORT_LOCATIONS"), $link);
+        $bar->appendButton('Link', 'download', JText::_('COM_CROWDFUNDING_EXPORT_LOCATIONS'), $link);
 
         $link = JRoute::_('index.php?option=com_crowdfunding&task=export.download&format=raw&type=states');
-        $bar->appendButton('Link', 'download', JText::_("COM_CROWDFUNDING_EXPORT_STATES"), $link);
+        $bar->appendButton('Link', 'download', JText::_('COM_CROWDFUNDING_EXPORT_STATES'), $link);
 
         JToolbarHelper::divider();
-        JToolbarHelper::publishList("locations.publish");
-        JToolbarHelper::unpublishList("locations.unpublish");
+        JToolbarHelper::publishList('locations.publish');
+        JToolbarHelper::unpublishList('locations.unpublish');
         JToolbarHelper::divider();
-        JToolbarHelper::deleteList(JText::_("COM_CROWDFUNDING_DELETE_ITEMS_QUESTION"), "locations.delete");
+        JToolbarHelper::deleteList(JText::_('COM_CROWDFUNDING_DELETE_ITEMS_QUESTION'), 'locations.delete');
         JToolbarHelper::divider();
-        JToolbarHelper::custom('locations.backToDashboard', "dashboard", "", JText::_("COM_CROWDFUNDING_DASHBOARD"), false);
+        JToolbarHelper::custom('locations.backToDashboard', 'dashboard', '', JText::_('COM_CROWDFUNDING_DASHBOARD'), false);
     }
 
     /**
@@ -155,6 +150,6 @@ class CrowdfundingViewLocations extends JViewLegacy
 
         JHtml::_('formbehavior.chosen', 'select');
 
-        JHtml::_('prism.ui.joomlaList');
+        JHtml::_('Prism.ui.joomlaList');
     }
 }